+++ src/app/_layout.tsx
@@ -195,21 +195,27 @@ const RootLayout = () => {
                             headerShown: false,
                           }}
                         />
-
                         <Stack.Screen
                           name="(public)/devices/[deviceId]"
                           options={{
                             headerShown: false,
-                            // presentation: 'modal',
-                            // ...(Platform.OS === 'android' && { animation: 'slide_from_bottom' }),
                           }}
                         />
                         <Stack.Screen
                           name="(public)/devices/new"
                           options={{
                             headerShown: false,
-                            // presentation: 'modal',
-                            // ...(Platform.OS === 'android' && { animation: 'slide_from_bottom' }),
+                          }}
+                        />
+
+                        <Stack.Screen
+                          name="(public)/changes"
+                          options={{
+                            headerShown: false,
+                            presentation: 'modal',
+                            ...(Platform.OS === 'android' && {
+                              animation: 'slide_from_bottom',
+                            }),
                           }}
                         />
                       </Stack>