+++ CHANGELOG.md
@@ -5,13 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## unreleased
+## [unreleased]
 
 ### Added
 
 - Surprise 🤫
 - Devices management.
 - Flex desks on `P'ti Poulailler` floor plan.
+- Releases notes on `about` screen.
+- Social networks bottom sheet on `settings` screen.
 
 ### Changed